refactor(arch): Phase 3 — PR #484 리뷰 4차 수정
- B1: FmpFundamentalClient → shared/api/fmp/fundamentalClient.ts 이동 (cross-entity 3/4 해소)
- 잔여 2건(options-chain) Phase 9 해소 주석 유지
- entities/fundamental/index.ts → shared re-export placeholder
- 테스트 colocation 정리
