+++ src/entities/fundamental/index.ts
@@ -1,1 +1,4 @@
-export { FmpFundamentalClient, type FmpEarningsReportItem } from './api';
+export {
+    FmpFundamentalClient,
+    type FmpEarningsReportItem,
+} from '@/shared/api/fmp/fundamentalClient';
