+++ src/shared/api/fmp/__tests__/fundamentalClient.test.ts
@@ -1,4 +1,4 @@
-import { FmpFundamentalClient } from '../api';
+import { FmpFundamentalClient } from '../fundamentalClient';
 
 const mockFetch = jest.fn();
